requirements.txt
asgiref==3.10.0
attrs==25.4.0
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Pillow==11.0.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
whitenoise==6.6.0
django-import-export
